+++ pyinaturalist/models/__init__.py
@@ -92,7 +92,14 @@ from pyinaturalist.models.project import (
     ProjectObservationField,
     ProjectUser,
 )
-from pyinaturalist.models.observation import Observation, Observations, QualityMetric, Fave, Vote
+from pyinaturalist.models.observation import (
+    Observation,
+    Observations,
+    Application,
+    Fave,
+    QualityMetric,
+    Vote,
+)
 from pyinaturalist.models.search import SearchResult