+++ HISTORY.md
@@ -20,6 +20,7 @@ Add support for searching observations by observation fields, using a new `obser
 * Add `Taxon.flatten()` method to return a taxon and its descendants as a flat list
 * Add `Observation.ident_taxon_ids` dynamic property to get all identification taxon IDs (with ancestors)
 * Add `Observation.cumulative_ids` dynamic property to calculate agreements/total community identifications
+* Add `Application` model for `Observation.application` (for v2 API)
 * Add `Fave` model for `Observation.faves`
 * Add `QualityMetric` model for `Observation.quality_metrics`
 * Add `Sound` model for `Observation.sounds`